Split off statement part from SPCG
